@@ -8,25 +8,29 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.3.0",
+        "php": "^7.3|^8.0",
         "ext-exif": "*",
         "ext-gd": "*",
         "facade/ignition": "^2.3.6",
-        "fideloper/proxy": "^4.2",
+        "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
-        "fzaninotto/faker": "^1.9.1",
+        "fakerphp/faker": "^1.9.1",
         "guzzlehttp/guzzle": "^7.0.1",
-        "inertiajs/inertia-laravel": "^0.2",
-        "laravel/framework": "^8.0",
+        "inertiajs/inertia-laravel": "^0.4",
+        "laravel/framework": "^8.40",
         "laravel/legacy-factories": "^1.0",
-        "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "league/glide": "2.0.x-dev",
-        "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^5.0",
-        "phpunit/phpunit": "^9.3",
+        "laravel/tinker": "^2.5",
+        "laravel/ui": "^3.0",
+        "league/glide-laravel": "^1.0",
         "reinink/remember-query-strings": "^0.1.0",
         "tightenco/ziggy": "^0.8.0"
+    },
+    "require-dev": {
+      "facade/ignition": "^2.5",
+      "laravel/sail": "^1.0.1",
+      "mockery/mockery": "^1.4.2",
+      "nunomaduro/collision": "^5.0",
+      "phpunit/phpunit": "^9.3.3"
     },
     "config": {
         "optimize-autoloader": true,
@@ -40,12 +44,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
