@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Models\Account;
 use App\Models\Contact;
 use App\Models\Organization;
@@ -8,11 +10,16 @@ use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
 {
+    /**
+     * Seed the application's database.
+     *
+     * @return void
+     */
     public function run()
     {
         $account = Account::create(['name' => 'Acme Corporation']);
 
-        factory(User::class)->create([
+        User::factory()->create([
             'account_id' => $account->id,
             'first_name' => 'John',
             'last_name' => 'Doe',
@@ -20,12 +27,12 @@ class DatabaseSeeder extends Seeder
             'owner' => true,
         ]);
 
-        factory(User::class, 5)->create(['account_id' => $account->id]);
+        User::factory(5)->create(['account_id' => $account->id]);
 
-        $organizations = factory(Organization::class, 100)
+        $organizations = Organization::factory(100)
             ->create(['account_id' => $account->id]);
 
-        factory(Contact::class, 100)
+        Contact::factory(100)
             ->create(['account_id' => $account->id])
             ->each(function ($contact) use ($organizations) {
                 $contact->update(['organization_id' => $organizations->random()->id]);
