@@ -3,7 +3,6 @@
 namespace Tests\Feature;
 
 use App\Models\Account;
-use App\Models\Contact;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
@@ -16,77 +15,118 @@ class ContactsTest extends TestCase
     {
         parent::setUp();
 
-        $account = Account::create(['name' => 'Acme Corporation']);
-
-        $this->user = factory(User::class)->create([
-            'account_id' => $account->id,
+        $this->user = User::factory()->create([
+            'account_id' => Account::create(['name' => 'Acme Corporation'])->id,
             'first_name' => 'John',
             'last_name' => 'Doe',
             'email' => 'johndoe@example.com',
             'owner' => true,
         ]);
+
+        $organization = $this->user->account->organizations()->create(['name' => 'Example Organization Inc.']);
+
+        $this->user->account->contacts()->createMany([
+            [
+                'organization_id' => $organization->id,
+                'first_name' => 'Martin',
+                'last_name' => 'Abbott',
+                'email' => 'martin.abbott@example.com',
+                'phone' => '555-111-2222',
+                'address' => '330 Glenda Shore',
+                'city' => 'Murphyland',
+                'region' => 'Tennessee',
+                'country' => 'US',
+                'postal_code' => '57851',
+            ], [
+                'organization_id' => $organization->id,
+                'first_name' => 'Lynn',
+                'last_name' => 'Kub',
+                'email' => 'lynn.kub@example.com',
+                'phone' => '555-333-4444',
+                'address' => '199 Connelly Turnpike',
+                'city' => 'Woodstock',
+                'region' => 'Colorado',
+                'country' => 'US',
+                'postal_code' => '11623',
+            ],
+        ]);
     }
 
     public function test_can_view_contacts()
     {
-        $this->user->account->contacts()->saveMany(
-            factory(Contact::class, 5)->make()
-        );
-
         $this->actingAs($this->user)
             ->get('/contacts')
-            ->assertStatus(200)
-            ->assertPropCount('contacts.data', 5)
-            ->assertPropValue('contacts.data', function ($contacts) {
-                $this->assertEquals(
-                    ['id', 'name', 'phone', 'city',
-                    'deleted_at', 'organization'],
-                    array_keys($contacts[0])
-                );
-            });
+            ->assertInertia(fn ($assert) => $assert
+                ->component('Contacts/Index')
+                ->has('contacts.data', 2)
+                ->has('contacts.data.0', fn ($assert) => $assert
+                    ->where('id', 1)
+                    ->where('name', 'Martin Abbott')
+                    ->where('phone', '555-111-2222')
+                    ->where('city', 'Murphyland')
+                    ->where('deleted_at', null)
+                    ->has('organization', fn ($assert) => $assert
+                        ->where('name', 'Example Organization Inc.')
+                    )
+                )
+                ->has('contacts.data.1', fn ($assert) => $assert
+                    ->where('id', 2)
+                    ->where('name', 'Lynn Kub')
+                    ->where('phone', '555-333-4444')
+                    ->where('city', 'Woodstock')
+                    ->where('deleted_at', null)
+                    ->has('organization', fn ($assert) => $assert
+                        ->where('name', 'Example Organization Inc.')
+                    )
+                )
+            );
     }
 
     public function test_can_search_for_contacts()
     {
-        $this->user->account->contacts()->saveMany(
-            factory(contact::class, 5)->make()
-        )->first()->update([
-            'first_name' => 'Greg',
-            'last_name' => 'Andersson'
-        ]);
-
         $this->actingAs($this->user)
-            ->get('/contacts?search=Greg')
-            ->assertStatus(200)
-            ->assertPropValue('filters.search', 'Greg')
-            ->assertPropCount('contacts.data', 1)
-            ->assertPropValue('contacts.data', function ($contacts) {
-                $this->assertEquals('Greg Andersson', $contacts[0]['name']);
-            });
+            ->get('/contacts?search=Martin')
+            ->assertInertia(fn ($assert) => $assert
+                ->component('Contacts/Index')
+                ->where('filters.search', 'Martin')
+                ->has('contacts.data', 1)
+                ->has('contacts.data.0', fn ($assert) => $assert
+                    ->where('id', 1)
+                    ->where('name', 'Martin Abbott')
+                    ->where('phone', '555-111-2222')
+                    ->where('city', 'Murphyland')
+                    ->where('deleted_at', null)
+                    ->has('organization', fn ($assert) => $assert
+                        ->where('name', 'Example Organization Inc.')
+                    )
+                )
+            );
     }
 
     public function test_cannot_view_deleted_contacts()
     {
-        $this->user->account->contacts()->saveMany(
-            factory(contact::class, 5)->make()
-        )->first()->delete();
+        $this->user->account->contacts()->firstWhere('first_name', 'Martin')->delete();
 
         $this->actingAs($this->user)
             ->get('/contacts')
-            ->assertStatus(200)
-            ->assertPropCount('contacts.data', 4);
+            ->assertInertia(fn ($assert) => $assert
+                ->component('Contacts/Index')
+                ->has('contacts.data', 1)
+                ->where('contacts.data.0.name', 'Lynn Kub')
+            );
     }
 
     public function test_can_filter_to_view_deleted_contacts()
     {
-        $this->user->account->contacts()->saveMany(
-            factory(contact::class, 5)->make()
-        )->first()->delete();
+        $this->user->account->contacts()->firstWhere('first_name', 'Martin')->delete();
 
         $this->actingAs($this->user)
             ->get('/contacts?trashed=with')
-            ->assertStatus(200)
-            ->assertPropValue('filters.trashed', 'with')
-            ->assertPropCount('contacts.data', 5);
+            ->assertInertia(fn ($assert) => $assert
+                ->component('Contacts/Index')
+                ->has('contacts.data', 2)
+                ->where('contacts.data.0.name', 'Martin Abbott')
+                ->where('contacts.data.1.name', 'Lynn Kub')
+            );
     }
 }
