@@ -2,12 +2,22 @@
 
 namespace App\Http\Controllers;
 
-use League\Glide\Server;
+use Illuminate\Http\Request;
+use League\Glide\ServerFactory;
+use Illuminate\Contracts\Filesystem\Filesystem;
+use League\Glide\Responses\LaravelResponseFactory;
 
 class ImagesController extends Controller
 {
-    public function show(Server $glide)
+    public function show(Filesystem $filesystem, Request $request, $path)
     {
-        return $glide->fromRequest()->response();
+        $server = ServerFactory::create([
+            'response' => new LaravelResponseFactory($request),
+            'source' => $filesystem->getDriver(),
+            'cache' => $filesystem->getDriver(),
+            'cache_path_prefix' => '.glide-cache',
+        ]);
+
+        return $server->getImageResponse($path, $request->all());
     }
 }
