@@ -17,8 +17,15 @@ class OrganizationsController extends Controller
             'organizations' => Auth::user()->account->organizations()
                 ->orderBy('name')
                 ->filter(Request::only('search', 'trashed'))
-                ->paginate()
-                ->only('id', 'name', 'phone', 'city', 'deleted_at'),
+                ->paginate(10)
+                ->withQueryString()
+                ->through(fn ($organization) => [
+                    'id' => $organization->id,
+                    'name' => $organization->name,
+                    'phone' => $organization->phone,
+                    'city' => $organization->city,
+                    'deleted_at' => $organization->deleted_at,
+                ]),
         ]);
     }
 
