@@ -1,16 +1,35 @@
 <?php
 
-use Faker\Generator as Faker;
+namespace Database\Factories;
 
-$factory->define(App\Models\Organization::class, function (Faker $faker) {
-    return [
-        'name' => $faker->company,
-        'email' => $faker->companyEmail,
-        'phone' => $faker->tollFreePhoneNumber,
-        'address' => $faker->streetAddress,
-        'city' => $faker->city,
-        'region' => $faker->state,
-        'country' => 'US',
-        'postal_code' => $faker->postcode,
-    ];
-});
+use App\Models\Organization;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class OrganizationFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Organization::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->company,
+            'email' => $this->faker->companyEmail,
+            'phone' => $this->faker->tollFreePhoneNumber,
+            'address' => $this->faker->streetAddress,
+            'city' => $this->faker->city,
+            'region' => $this->faker->state,
+            'country' => 'US',
+            'postal_code' => $this->faker->postcode,
+        ];
+    }
+}
