@@ -13,7 +13,9 @@ return [
     |
     */
 
-    'previous' => '&laquo; Previous',
-    'next' => 'Next &raquo;',
+    // 'previous' => '&laquo; Previous',
+    // 'next' => 'Next &raquo;',
+    'previous' => 'Previous',
+    'next' => 'Next',
 
 ];
