@@ -1,9 +1,17 @@
 <!DOCTYPE html>
-<html class="h-full bg-gray-200">
+<html class="h-full bg-gray-100">
 <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0">
-    <link href="{{ mix('/css/app.css') }}" rel="stylesheet">
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="icon" type="image/svg+xml" href="/favicon.svg">
+    <link href="{{ asset('css/app.css') }}" rel="stylesheet">
+
+    {{-- Inertia --}}
+    <script src="https://polyfill.io/v3/polyfill.min.js?features=smoothscroll,NodeList.prototype.forEach,Promise,Object.values,Object.assign" defer></script>
+
+    {{-- Ping CRM --}}
+    <script src="https://polyfill.io/v3/polyfill.min.js?features=String.prototype.startsWith" defer></script>
+
     <script src="{{ mix('/js/app.js') }}" defer></script>
     @routes
 </head>
