@@ -1,8 +1,7 @@
+const path = require('path')
+const mix = require('laravel-mix')
 const cssImport = require('postcss-import')
 const cssNesting = require('postcss-nesting')
-const mix = require('laravel-mix')
-const path = require('path')
-const tailwindcss = require('tailwindcss')
 
 /*
  |--------------------------------------------------------------------------
@@ -21,7 +20,7 @@ mix.js('resources/js/app.js', 'public/js')
     postCss: [
       cssImport(),
       cssNesting(),
-      tailwindcss('tailwind.config.js'),
+      require('tailwindcss'),
     ],
   })
   .webpackConfig({
